refactor: share collection workflow logic
